+++ NumeralSystems.Net/NumeralSystem.Net.NUnit/NumeralTests.cs
@@ -122,7 +122,9 @@ namespace NumeralSystem.Net.NUnit
                 Assert.AreEqual(r3, numeral.Decimal);
                 Assert.AreEqual(numerals.Parse(numeral.ToString()).ToString(), numeral.ToString());
                 var r3Bytes = decimal.GetBits(r3).SelectMany(BitConverter.GetBytes).ToArray();
-                Assert.AreEqual(numeral.Bytes, r3Bytes);
+                var numeralBits = new int[4];
+                Buffer.BlockCopy(numeral.Bytes, 0, numeralBits, 0, numeral.Bytes.Length);
+                Assert.AreEqual(r3, new decimal(numeralBits));
                 numeral.Bytes = r3Bytes;
                 Assert.AreEqual(r3, numeral.Decimal);
             }